% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa_mass_data}
\alias{aa_mass_data}
\title{Amino acid masses}
\format{
A data frame with 22 rows and 5 variables: \describe{
\item{amino.acid}{amino acid name}
\item{X1.letter.code}{amino acid in 1-letter code format}
\item{X3.letter.code}{amino acid in 3-letter code format}
\item{monoisotopic}{monoisotopic mass, Da}
\item{average}{average mass, Da}
}
}
\source{
\url{https://web.expasy.org/findmod/findmod_masses.html#AA}
}
\usage{
aa_mass_data
}
\description{
A dataset containing the masses of all amino acids.
}
\keyword{datasets}
