% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fppercell.R
\name{calc_fppercell}
\alias{calc_fppercell}
\title{Calculate FP per cell values}
\usage{
calc_fppercell(
  data_csv,
  timecourse = TRUE,
  flu_channels,
  flu_labels,
  remove_wells,
  get_rfu_od = TRUE,
  get_mol_cell = FALSE,
  plate_type = 96,
  outfolder = "."
)
}
\arguments{
\item{data_csv}{path to a CSV file containing processed plate reader data}

\item{timecourse}{logical. Is the data timecourse/kinetic data and does it
include a variable called 'time'?}

\item{flu_channels}{the column names for the NORMALISED fluorescence data}

\item{flu_labels}{the column names for the CALIBRATED fluorescence data}

\item{remove_wells}{list of coordinates of wells to be removed from analysis
(e.g. empty wells)}

\item{get_rfu_od}{logical. if TRUE, uses \code{normalised_FP} and \code{normalised_OD}
to calculate FP per cell as FP/OD in relative fluorescence units/relative
OD (RFU/OD).}

\item{get_mol_cell}{logical. if TRUE, uses \code{calibrated_FP} and
\code{calibrated_OD} to calculate FP per cell as FP/OD in relative fluorescence
units/relative OD (molecules/cell).}

\item{plate_type}{type of plate. numeric, i.e. '96' for 96-well plate.
Defines the rows and columns used for plotting figures. Defaults to '96'.}

\item{outfolder}{path to folder where output files should be saved. Defaults
to current working directory.}
}
\value{
a data.frame with columns for each FP/cell calculation
}
\description{
Takes as input timecourse plate reader data processed with \code{process_plate}
and uses normalised/calibrated values to calculate per-cell values. Adds
column(s) for FP/cell as either: (a) \code{normalisedFP/normalisedOD} (RFU/OD),
(b) \code{calibratedFP/calibratedOD} (molecules/cell). Plots results and returns a
dataframe. Note that technically, units of molecules are 'molecules of
equivalent FP' and cells are 'particles of equivalent microspheres'.
}
\examples{
\dontrun{
  pc_data_mCherry <- calc_fppercell(
    data_csv = "mcherry_parsed_processed.csv",
    flu_channels = c("red1red1"), flu_labels = c("mCherry"),
    remove_wells = c("A11"),
    get_rfu_od = TRUE, get_mol_cell = TRUE,
    outfolder = file.path("plots")
  )
}
}
