% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_od.R
\name{calibrate_od}
\alias{calibrate_od}
\title{Convert arbitrary optical density units to calibrated units}
\usage{
calibrate_od(pr_data, od_name, instr, conversion_factors_csv)
}
\arguments{
\item{pr_data}{a dataframe of parsed plate reader data}

\item{od_name}{the column name for the optical density data}

\item{instr}{character string to represent instrument. If do_calibrate =
TRUE, used for filtering \code{od_coeffs_csv} and \code{fluor_coeffs_csv} files for
conversion factors of the relevant instrument.}

\item{conversion_factors_csv}{path of the CSV file containing
conversion factors for optical density}
}
\value{
an updated data.frame with an additional column for calibrated OD
}
\description{
Used by \code{process_plate} function for optical density calibration. Function
adds calibrated_OD column to the data, which is returned. Originally based on
\code{flopr::calibrate_od}, except \code{instr} argument added to allow selection of
conversion factor from table that may include multiple instruments, and
includes error checks that report to the user if conversion factors are
missing.
}
\keyword{internal}
