% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_quench_data}
\alias{cell_quench_data}
\title{cell quench data}
\format{
A data frame with 34 rows and 12 variables: \describe{
\item{channel_name}{name of filter set used}
\item{channel_ex}{excitation filter}
\item{channel_em}{emission filter}
\item{media}{media}
\item{calibrant}{calibrant}
\item{protein}{protein}
\item{cells_od600}{accurate OD600 cm-1 values}
\item{cells_od700}{accurate OD700 cm-1 values}
\item{volume}{volume in ul}
\item{measure}{filter set plus gain}
\item{gain}{gain}
\item{autof_norm_value}{fluorescence values normalised to remove the contribution of cellular autofluorescence}
\item{norm_fc_value}{fold change between the fluorescence sample and its relevant negative control without cells}
}
}
\usage{
cell_quench_data
}
\description{
A table containing experimental data for informing how cells quench fluorescence of E. coli cells.
}
\keyword{datasets}
