% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_flu.R
\name{correct_flu}
\alias{correct_flu}
\title{Correct normalised fluorescence to compensate for cellular quenching effects}
\usage{
correct_flu(pr_data, od_type, flu_channel)
}
\arguments{
\item{pr_data}{a data.frame of parsed plate reader data}

\item{od_type}{Which OD-type was used? Required for quench correction.
"OD600" or "OD700".}

\item{flu_channel}{fluorescent channel name}
}
\value{
an updated data.frame with an additional column for corrected fluorescence
}
\description{
Used by \code{process_plate} function for cell quench correction. The presence of
cells, particularly at high densities, can 'quench' fluorescence such that
fluorescence readings are lower than expected from the equivalent FP
concentration in vitro. This function compensates for this effect using
empirical data that relates the fold decrease in fluorescence expected, given
a certain cell density.
}
