% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_format_functions.R
\name{find_columns}
\alias{find_columns}
\title{Find columns of a given plate type}
\usage{
find_columns(plate_type = 96)
}
\arguments{
\item{plate_type}{type of plate. numeric, ie. \code{96} for 96-well plate.}
}
\description{
Internal function.
}
