% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_format_functions.R
\name{find_plate_format}
\alias{find_plate_format}
\title{Find row and column configuration of a given plate type}
\usage{
find_plate_format(plate_type)
}
\arguments{
\item{plate_type}{type of plate. numeric, i.e. \code{96} for 96-well plate.}
}
\description{
Internal function.
}
\keyword{internal}
