% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_format_functions.R
\name{find_wells}
\alias{find_wells}
\title{Find wells of a given plate format}
\usage{
find_wells(plate_type = 96)
}
\arguments{
\item{plate_type}{type of plate. numeric, i.e. \code{96} for 96-well plate.}
}
\value{
a list containing the names of all the wells of a specified plate
}
\description{
Works out all wells in that plate type and returns as a list.
}
\examples{
  find_wells(96)
  find_wells(24)

}
