% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_format_functions.R
\name{find_wells_in_column}
\alias{find_wells_in_column}
\title{List of all wells in a specific column of a 96-well plate}
\usage{
find_wells_in_column(columns, plate_type = 96)
}
\arguments{
\item{columns}{number or list of numbers, representing column(s) such as \code{1} or \code{12}. character strings also accepted.}

\item{plate_type}{type of plate. numeric, ie. \code{96} for 96-well plate.}
}
\value{
a list containing the names of all the wells in the specified column(s) of a specified plate
}
\description{
List of all wells in a specific column of a 96-well plate
}
\examples{
 find_wells_in_column("1")
 find_wells_in_column("12")
 find_wells_in_column(1)
 find_wells_in_column(c(1,3), plate_type = 6)
}
