% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_format_functions.R
\name{find_wells_in_row}
\alias{find_wells_in_row}
\title{List of all wells in a specific row of a given plate type}
\usage{
find_wells_in_row(rows, plate_type = 96)
}
\arguments{
\item{rows}{character string, or list of character strings, representing row(s) such as "A" or "H"}

\item{plate_type}{type of plate. numeric, i.e. \code{96} for 96-well plate.}
}
\value{
a list containing the names of all the wells in the specified row(s) of a specified plate
}
\description{
List of all wells in a specific row of a given plate type
}
\examples{
 find_wells_in_row("A")
 find_wells_in_row(c("A","D"), plate_type = 24)
}
