% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flu_norm.R
\name{flu_norm}
\alias{flu_norm}
\title{Normalise fluorescence against negative well autofluorescence}
\usage{
flu_norm(
  pr_data,
  neg_well,
  blank_well,
  flu_name,
  af_model,
  data_csv,
  outfolder
)
}
\arguments{
\item{pr_data}{a long data.frame containing you plate reader data with OD
normalised}

\item{neg_well}{the well coordinates of a non-fluorescent control}

\item{blank_well}{the well coordinates of a media blank}

\item{flu_name}{the column name of the fluorescence chanel to normalise}

\item{af_model}{model used to fit negative control autofluorescence.
For now these include "polynomial", "invers_poly", "exponential", "spline" or "loess".
If set to NULL, no model is used, and fluorescence is normalised akin to OD: by subtracting the value for the blanks.}

\item{data_csv}{path to the original data. Used for saving normalisation curve plots.}

\item{outfolder}{path to folder where output files should be saved. Defaults
to current working directory.}
}
\value{

}
\description{
Used by \code{process_plate} function for fluorescence normalisation. Remains
virtually unchanged from \code{flopr::flu_norm}, except that \code{af_model} can be set
to \code{NULL} to normalise to blank wells instead, and files are saved to a
specified \code{outfolder}.
}
