% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fpbase_data}
\alias{fpbase_data}
\title{fluorescent protein properties from fpbase}
\format{
A data frame with 593 rows and 16 variables: \describe{
\item{url}{FPbase url} \item{name}{name} \item{stokes}{Stokes shift (nm)}
\item{slug}{slug - lower case name used in url} \item{ipg_id}{Identical
Protein Group ID at Pubmed} \item{agg}{oligomerisation state}
\item{ex_max}{wavelength of excitation peak (nm)} \item{em_max}{wavelength
of emission peak (nm)} \item{ext_coeff}{excitation coefficient (EC,
M-1cm-1)} \item{qy}{quantum yield} \item{pka}{pKa}
\item{brightness}{brightness} \item{bleach}{rate of
photobleaching/photostability (s)} \item{maturation}{maturation half time
(min)} \item{lifetime}{fluorescence lifetime (ns)} \item{cofactor}{cofactor
required for fluorescence (if any)}}
}
\source{
\url{https://www.fpbase.org/api/proteins/basic/?format=json}
}
\usage{
fpbase_data
}
\description{
A dataset containing some of the properties of all the 'basic' (single
fluorescent state) fluorescent proteins in the fpbase database (downloaded
2022.10.04).
}
\keyword{datasets}
