% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mw.R
\name{get_mw}
\alias{get_mw}
\title{Get molecular weight of a protein}
\usage{
get_mw(protein)
}
\arguments{
\item{protein}{character string of protein sequence using the 1-letter code.}
}
\value{
numeric molecular weight
}
\description{
Get the molecular weight of a protein in Daltons from its sequence alone.
Uses calculation (\url{https://web.expasy.org/compute_pi/pi_tool-doc.html}) and
amino acid mass data (\code{aa_mass_data},
\url{https://web.expasy.org/findmod/findmod_masses.html#AA}) from the Swiss
Bioinformatics Resource Portal, Expasy.
}
