% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_properties.R
\name{get_properties}
\alias{get_properties}
\title{Get properties of an FP from FPbase}
\usage{
get_properties(
  slug,
  verbose = TRUE,
  save_file = FALSE,
  outfolder = ".",
  filename = "fp_properties.csv"
)
}
\arguments{
\item{slug}{Name of FP, which is used to find the FP-relevant data lines in
the FPbase datasets. This argument is called \code{slug} because for FPbase
retrievals, this needs to be an exact match to the slug specified by
FPbase. (To find this, navigate to the FPbase entry for your FP and copy
the part after \verb{https://www.fpbase.org/protein/} (without the trailing
\code{/}).}

\item{verbose}{logical. Should the function print messages to allow the user
to check its progress? Defaults to TRUE.}

\item{save_file}{logical. Should the function save the output as a csv file?
Defaults to FALSE.}

\item{outfolder}{path to folder where output files should be saved. Defaults
to current working directory.}

\item{filename}{How to name the output csv file. Requires ".csv" at the end.
Defaults to "fp_properties.csv".}
}
\description{
Gets properties of an FP from FPbase, including the maximal excitation
wavelength and extinction coefficient, and optionally, saves these as csv.
}
\examples{
get_properties("mcherry")
}
