% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magellan_spectrum_parse.R
\name{magellan_spectrum_parse}
\alias{magellan_spectrum_parse}
\title{Parser for absorbance spectrum data}
\usage{
magellan_spectrum_parse(
  data_csv,
  layout_csv,
  wellstart = "A1",
  wellend = "H12"
)
}
\arguments{
\item{data_csv}{path to csv file from Tecan Spark plate reader}

\item{layout_csv}{path to csv file containing plate layout information}

\item{wellstart}{character string representing first well recorded in data file. Defaults to "A1".}

\item{wellend}{character string representing last well recorded in data file. Defaults to "H12".}
}
\value{
a data.frame containing the parsed plate reader data
}
\description{
Parser for absorbance spectrum data exported from a Tecan Spark plate reader using SparkControl Magellan software.
}
\examples{
parsed_data <- magellan_spectrum_parse(data_csv = "data/20210104_data.csv", layout_csv = "data/20210104_data_layout.csv", wellstart = "A1", wellend = "H12")
}
