% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_od.R
\name{norm_od}
\alias{norm_od}
\title{Normalise optical density against blank wells}
\usage{
norm_od(pr_data, blank_well, od_name, timecourse)
}
\arguments{
\item{pr_data}{a long data.frame containing your plate reader data}

\item{blank_well}{the well coordinates of one or more media blanks}

\item{od_name}{the column name for the optical density data}

\item{timecourse}{logical. Is the data timecourse/kinetic data and does it
include a variable called 'time'?}
}
\value{
an updated data.frame with an additional column "normalised_OD"
}
\description{
Used by \code{process_plate} function for OD normalisation. Remains virtually
unchanged from \code{flopr::od_norm}.
}
\keyword{internal}
