% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magellan_spectrum_parse.R
\name{parse_magellan_spectrum}
\alias{parse_magellan_spectrum}
\title{Parse Tecan Spark/Magellan data from absorbance spectrum data}
\usage{
parse_magellan_spectrum(
  data_csv,
  layout_csv,
  wellstart = "A1",
  wellend = "H12"
)
}
\arguments{
\item{data_csv}{path to CSV file from Tecan Spark plate reader}

\item{layout_csv}{path to CSV file containing plate layout information}

\item{wellstart}{character string representing first well recorded in data
file. Defaults to "A1".}

\item{wellend}{character string representing last well recorded in data file.
Defaults to "H12".}
}
\value{
a data.frame containing the parsed plate reader data
}
\description{
Parses absorbance spectrum data exported from a Tecan Spark plate reader
using Magellan software. Parsing consists of data extraction, data tidying,
and data joining to relevant 'plate layout' metadata.
}
\examples{
\dontrun{
  parsed_data <- parse_magellan_spectrum(
    data_csv = "data/20210104_data.csv",
    layout_csv = "data/20210104_data_layout.csv",
    wellstart = "A1", wellend = "H12"
  )
}
}
