% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sparkcontrol.R
\name{parse_sparkcontrol}
\alias{parse_sparkcontrol}
\title{Parse Tecan Spark/SparkControl data from endpoint and kinetic assays}
\usage{
parse_sparkcontrol(data_csv, metadata_csv, timeseries = FALSE)
}
\arguments{
\item{data_csv}{path to CSV file from Tecan Spark plate reader}

\item{metadata_csv}{path to CSV file containing metadata}

\item{timeseries}{logical. Is the data a timeseries? Defaults to FALSE.}
}
\value{
a data.frame containing the parsed plate reader data
}
\description{
Parses data exported from a Tecan Spark plate reader using SparkControl
software. Handles standard (endpoint) or timecourse (kinetic) data containing
absorbance and/or fluorescence readings, but cannot handle spectra, such as
absorbance scans. Parsing consists of data extraction, data tidying, and data
joining to relevant metadata. Originally based on \code{flopr::spark_parse()}
function.
}
\examples{
\dontrun{
  parsed_calib_plate <- parse_sparkcontrol(
    data_csv = "calibrations/20210104_calibration_data.csv",
    metadata_csv = "calibrations/20210104_calibration_metadata.csv",
    timeseries = FALSE
  )

  parsed_data <- parse_sparkcontrol(
    data_csv = "data/20210104_data.csv",
    metadata_csv = "data/20210104_metadata.csv",
    timeseries = TRUE
  )
}
}
