% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sparkcontrol_spectrum.R
\name{parse_sparkcontrol_spectrum}
\alias{parse_sparkcontrol_spectrum}
\title{Parse Tecan Spark/SparkControl data from absorbance spectrum data}
\usage{
parse_sparkcontrol_spectrum(
  data_csv,
  metadata_csv,
  wellstart = "A1",
  wellend = "H12"
)
}
\arguments{
\item{data_csv}{path to CSV file from Tecan Spark plate reader}

\item{metadata_csv}{path to CSV file containing metadata}

\item{wellstart}{character string representing first well recorded in data
file. Defaults to "A1".}

\item{wellend}{character string representing last well recorded in data file.
Defaults to "H12".}
}
\value{
a data.frame containing the parsed plate reader data
}
\description{
Parses absorbance spectrum data exported from a Tecan Spark plate reader
using SparkControl software. Parsing consists of data extraction, data
tidying, and data joining to relevant metadata. Based on
\code{parse_magellan_spectrum()}.
}
\examples{
\dontrun{
  parsed_data <- parse_sparkcontrol_spectrum(
    data_csv = "data/20210104_data.csv",
    metadata_csv = "data/20210104_metadata.csv",
    wellstart = "A1", wellend = "H12"
  )
}
}
