% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathlength_water_data}
\alias{pathlength_water_data}
\title{pathlength vs volume data}
\format{
A data frame with 10 rows and 6 variables: \describe{
\item{volume}{volume in ul} \item{pathlength}{pathlength in cm}
\item{expt}{experiment number denoting experimental repeat} }
}
\usage{
pathlength_water_data
}
\description{
A dataset containing experimentally derived pathlengths for microplate wells
containing water at different volumes. Pathlengths were derived by using the
equation \code{pathlength (cm) = kfactor_well / kfactor_1cm}. \code{Kfactor_well} is
defined as A975-A900 value for each well (there were 4 replicates per volume
per experiment). \code{Kfactor_1cm} values were derived using the \code{get_kfactor()}
function for water at 26oC.
}
\keyword{datasets}
