% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absorbance_spectrum.R
\name{plot_absorbance_spectrum}
\alias{plot_absorbance_spectrum}
\title{Plot absorbance spectrum}
\usage{
plot_absorbance_spectrum(
  spectrum_csv,
  subset_rows = FALSE,
  rows_to_keep = c("C", "D"),
  columns_to_keep = c(1, 12),
  xrange = c(200, 1000),
  pl_method = "calc_blanks",
  buffer_used = "water",
  concentration_used = 0,
  temperature_used = 25,
  outfolder = "."
)
}
\arguments{
\item{spectrum_csv}{path of a CSV file of your spectrum data}

\item{subset_rows}{logical. should script take a subset of the rows (or whole
table)? Defaults to FALSE.}

\item{rows_to_keep}{character array. If \code{subset_rows} is TRUE, script will
choose rows to keep from this list. Defaults to \code{c("C","D")}.}

\item{columns_to_keep}{numeric array. If \code{subset_rows} is TRUE, script will
choose rows to keep from this list. Defaults to \code{c(1,12)}. Note that blanks
are required for normalisation, so if blanks are in column12, it is
necessary to include 12 in this list.}

\item{xrange}{numerical lower and upper bounds of wavelengths, between which
the is subset for plotting. This can be useful for clear plates, which have
high background <300nm, so can set the \code{xrange} as c(300,800) or similar.}

\item{pl_method}{string denoting which method of path length normalisation to
use. Options are \code{calc_each}, \code{calc_blanks} and \code{volume}. All three are
always calculated (and compared in one of the output plots), but only the
chosen method is used. The \code{calc_each} method uses the path length
calculated from each well separately using the A900 and A975 measurements
and the k-factor of the buffer used. The \code{calc_blanks} method uses the path
length calculated from the blanks for all wells. Finally, the \code{volume}
method uses the path length expected from the given volume, using internal
path length data measured with water via the internal
\code{fpcountR::get_pathlength()} function. Defaults to \code{calc_blanks}.}

\item{buffer_used}{string corresponding to buffer used for assay, for use in
path length determination. Must be in table used by
\code{fpcountR::get_kfactor()}. To look at table, use
\code{fpcountR::view_kfactors()}. Defaults to "water".}

\item{concentration_used}{numeric value corresponding to concentration of
buffer, for use in path length determination. Must be in the same units as
used by table in \code{fpcountR::get_kfactor()}. To look at table, use
\code{fpcountR::view_kfactors()}. Defaults to 0, which effectively means
water. Needs changing if using different buffer!}

\item{temperature_used}{numeric value corresponding to temperature of assay,
for use in path length determination. Defaults to 25.}

\item{outfolder}{path to folder where output files should be saved. Defaults
to current working directory.}
}
\description{
Processes \emph{parsed absorbance spectrum data}, collected as a \emph{dilution series}
measured with an \emph{A200-1000 absorbance spectrum}. Corrects raw data to path
length of 1cm by a user-defined method, and normalises to the blanks. Plots
spectra and returns processed data.
}
\examples{
\dontrun{
  spectrum <- plot_absorbance_spectrum(
    spectrum_csv = "spectrum_parsed.csv",
    subset_rows = TRUE, rows_to_keep = c("C","D"), columns_to_keep = c(1:12),
    pl_method = "calc_blanks",
    buffer_used = "TBS", concentration_used = 0.005, temperature_used = 30,
    xrange = c(250,1000),
    outfolder = "spectrum"
  )
}
}
