% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_seq_data}
\alias{protein_seq_data}
\title{protein sequence data}
\format{
A data frame with 4 rows and 3 variables: \describe{
\item{name}{protein name} \item{slug}{protein slug - lower case name used
in url of fpbase entry} \item{sequence}{primary sequence in 1-letter code}
}
}
\usage{
protein_seq_data
}
\description{
A table containing the protein sequences used in the FPCountR paper.
}
\keyword{datasets}
