% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_metadata_template.R
\name{save_metadata_template}
\alias{save_metadata_template}
\title{Save a template metadata file}
\usage{
save_metadata_template(data_type, plate_type = 96, outfolder = "")
}
\arguments{
\item{data_type}{type of data, from absorbance spectrum data for calibrations
("absspectrum"), fluorescence data for calibrations ("fluordata") and
experimental data ("exptdata").}

\item{plate_type}{type of plate. numeric, i.e. \code{96} for 96-well plate.}

\item{outfolder}{path to folder where output files should be saved.}
}
\description{
Save a CSV file containing the minimum columns required by the functions that
parse and process the given \code{data_type} in \code{fpcountr}. One of three data
types can be chosen: absorbance spectrum data for calibrations
("absspectrum"), fluorescence data for calibrations ("fluordata") and
experimental data ("exptdata") The \code{well} column is populated with all the
wells in a given plate type specified by \code{plate_type}. The file is saved to
the location specified by \code{outfolder}.
}
\examples{
\dontrun{
  save_metadata_template(data_type = "exptdata", plate_type = 96, outfolder = ".")
}
}
