% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_kfactors.R
\name{view_kfactors}
\alias{view_kfactors}
\title{View k-factors}
\usage{
view_kfactors()
}
\description{
Displays the data tables used for k-factor calculations in \code{get_kfactors()}, along with notes.
}
\examples{
view_kfactors()
}
