% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_flu.R
\name{calibrate_flu}
\alias{calibrate_flu}
\title{Convert arbitrary fluorescence units to calibrated units}
\usage{
calibrate_flu(
  pr_data,
  flu_instr,
  flu_channel,
  flu_gain,
  flu_slug,
  flu_label,
  do_quench_correction,
  conversion_factors_csv
)
}
\arguments{
\item{pr_data}{a data.frame of parsed plate reader data}

\item{flu_instr}{character string to represent instrument}

\item{flu_channel}{the column names for the fluorescence data}

\item{flu_gain}{gain of the fluorescence channel}

\item{flu_slug}{name of fluorescent protein in FPbase slug format}

\item{flu_label}{the column names to be given to
the calibration. May be identical to flu_slug or flu_channel, but
recommended is to make it obvious which FP is being calibrated, e.g.
"mCherry", as channel names may be non-specifically named e.g. "red1red1".}

\item{do_quench_correction}{logical. Should function correct for anticipated
quenching of fluorescence, depending on the cell density?}

\item{conversion_factors_csv}{path of the CSV file
containing predicted conversion factors for the fluorescent channels}
}
\value{
an updated data.frame with an additional column for calibrated fluorescence
}
\description{
Used by \code{process_plate} function for fluorescence calibration. Function adds
calibrated fluorescence column to the data, which is returned. Originally
based on \code{flopr::calibrate_flu}, but with multiple changes. A list of
arguments have been added to allow selection of required conversion factor
from table that may include conversion factors from multiple instruments,
FPs, etc., and function now includes error checks that report to the user if
conversion factors are missing.
}
\keyword{internal}
