% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extcoeff_a280.R
\name{get_extcoeff_a280}
\alias{get_extcoeff_a280}
\title{Get a protein's molar extinction coefficient at A280}
\usage{
get_extcoeff_a280(
  protein,
  disulphides = FALSE,
  showWarnings = TRUE,
  showMessages = TRUE,
  protein_name = "-",
  buffer = "-",
  mol_weight = NULL,
  save_file = FALSE,
  outfolder = ""
)
}
\arguments{
\item{protein}{character string of protein sequence using 1-letter code}

\item{disulphides}{logical. Does protein have disulphides?}

\item{showWarnings}{logical. Should function show warnings?}

\item{showMessages}{logical. Should function show messages?}

\item{protein_name}{character string of protein name. Optional.}

\item{buffer}{character string of buffer. Optional.}

\item{mol_weight}{numerical value for molecular weight (g/mol). Optional. If
specified the function gives extinction coefficients for 1\% (10mg/ml) and
0.1\% (1mg/ml) solutions too.}

\item{save_file}{logical. Should function save CSV file of output?}

\item{outfolder}{path to folder where output files should be saved.}
}
\value{
data frame of extinction coefficients
}
\description{
Work out the theoretical molar extinction coefficient of a protein at 280nm
(EC280, M-1cm-1) using only the protein's primary sequence, using the
ProtParam method (Pace values). A full explanation of the method can be found
at \url{https://web.expasy.org/protparam/protparam-doc.html}.
}
