% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mw.R
\name{get_mw}
\alias{get_mw}
\title{Get molecular weight of a protein}
\usage{
get_mw(protein)
}
\arguments{
\item{protein}{character string of protein sequence using the 1-letter code.}
}
\description{
Get molecular weight of protein in Daltons from its sequence alone.
}
