% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathlength.R
\name{get_pathlength}
\alias{get_pathlength}
\title{Get path length in a typical 96-well plate}
\usage{
get_pathlength(test_volume, save_file = FALSE, outfolder = "")
}
\arguments{
\item{test_volume}{numeric value of volume whose pathlength is required, in
microlitres (ul)}

\item{save_file}{logical. Should the function plot the model and prediction?
Defaults to FALSE.}

\item{outfolder}{path to folder where output files should be saved.}
}
\value{
numeric path length in cm
}
\description{
Calculates the path length of an input \code{test_volume} by comparing it to a
linear model using internal data of pathlengths vs volume. ... To work out
the pathlength in a microplate assay, there are two options. Option 1:
Measure the pathlength each time you do an assay by taking A975-A900 for each
well as the \code{kfactor(well)}, and calculating \code{pathlength} = \code{kfactor(well)} /
\verb{kfactor(1cm pathlength)}. The 1cm k-factors can be derived using
\code{get_kfactor()} function. Option 2 (used here): Create a standardised dataset
of pathlength vs volume by measuring the A975-A900 of defined volumes of
buffer, and calculating the pathlengths of each (the \code{pathlength_water_data}
dataset) and use this to create a linear model of pathlength vs volume.
Experimental data suggests that aqueous buffers give very similar pathlength
values for a given volume, so here we use data from water to approximate all
aqueous buffers.
}
\examples{
get_pathlength(200)
}
